Worked on map image issues.

+++ src/main/java/agh/BalloonType.java
@@ -8,11 +8,11 @@ public enum BalloonType {
     BLUE,
     GREEN;
 
-    public double getStartingSpeed() {
+    public int getStartingSpeed() {
         return switch (this) {
-            case RED -> 1.0;
-            case BLUE -> 2.0;
-            case GREEN -> 3.0;
+            case RED -> 1;
+            case BLUE -> 2;
+            case GREEN -> 3;
         };
     }
